scripts/requirements.txt
requests>=2.28.0
python-frontmatter>=1.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
